[API stabilization] Cbor
    * Use sealed-class scheme
    * Mover all internals to a separate file
    * Hide CborDecodingException
