+++ formats/cbor/api/kotlinx-serialization-cbor.api
@@ -1,21 +1,25 @@
-public final class kotlinx/serialization/cbor/Cbor : kotlinx/serialization/BinaryFormat {
+public abstract class kotlinx/serialization/cbor/Cbor : kotlinx/serialization/BinaryFormat {
 	public static final field Default Lkotlinx/serialization/cbor/Cbor$Default;
-	public fun <init> ()V
-	public fun <init> (ZLkotlinx/serialization/modules/SerializersModule;)V
-	public synthetic fun <init> (ZLkotlinx/serialization/modules/SerializersModule;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public synthetic fun <init> (ZLkotlinx/serialization/modules/SerializersModule;Ljava/lang/Void;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public fun decodeFromByteArray (Lkotlinx/serialization/DeserializationStrategy;[B)Ljava/lang/Object;
 	public fun encodeToByteArray (Lkotlinx/serialization/SerializationStrategy;Ljava/lang/Object;)[B
+	public fun getSerializersModule ()Lkotlinx/serialization/modules/SerializersModule;
+}
+
+public final class kotlinx/serialization/cbor/Cbor$Default : kotlinx/serialization/cbor/Cbor {
+}
+
+public final class kotlinx/serialization/cbor/CborBuilder {
 	public final fun getEncodeDefaults ()Z
-	public fun getSerializersModule ()Lkotlinx/serialization/modules/SerializersModule;
+	public final fun getSerializersModule ()Lkotlinx/serialization/modules/SerializersModule;
+	public final fun setEncodeDefaults (Z)V
+	public final fun setSerializersModule (Lkotlinx/serialization/modules/SerializersModule;)V
 }
 
-public final class kotlinx/serialization/cbor/Cbor$Default : kotlinx/serialization/BinaryFormat {
-	public fun decodeFromByteArray (Lkotlinx/serialization/DeserializationStrategy;[B)Ljava/lang/Object;
-	public fun encodeToByteArray (Lkotlinx/serialization/SerializationStrategy;Ljava/lang/Object;)[B
-	public fun getSerializersModule ()Lkotlinx/serialization/modules/SerializersModule;
-}
-
-public final class kotlinx/serialization/cbor/CborDecodingException : kotlinx/serialization/SerializationException {
-	public fun <init> (Ljava/lang/String;I)V
+public final class kotlinx/serialization/cbor/CborKt {
+	public static final fun Cbor (Lkotlinx/serialization/cbor/Cbor;Lkotlin/jvm/functions/Function1;)Lkotlinx/serialization/cbor/Cbor;
+	public static final fun Cbor (Lkotlinx/serialization/modules/SerializersModule;)Lkotlinx/serialization/cbor/Cbor;
+	public static final fun Cbor (ZLkotlinx/serialization/modules/SerializersModule;)Lkotlinx/serialization/cbor/Cbor;
+	public static synthetic fun Cbor$default (Lkotlinx/serialization/cbor/Cbor;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)Lkotlinx/serialization/cbor/Cbor;
 }
