+++ formats/cbor/commonMain/src/kotlinx/serialization/cbor/internal/CborDecodingException.kt
@@ -2,11 +2,11 @@
  * Copyright 2017-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package kotlinx.serialization.cbor
+package kotlinx.serialization.cbor.internal
 
 import kotlinx.serialization.*
 
-public class CborDecodingException(expected: String, foundByte: Int) :
+internal class CborDecodingException(expected: String, foundByte: Int) :
     SerializationException("Expected $expected, but found ${printByte(foundByte)}")
 
 internal fun printByte(b: Int): String {
