+++ formats/cbor/commonTest/src/kotlinx/serialization/cbor/CborPolymorphismTest.kt
@@ -14,7 +14,7 @@ class CborPolymorphismTest {
         data class B(val b: String) : A()
     }
 
-    val cbor = Cbor(serializersModule = SimplePolymorphicModule)
+    val cbor = Cbor { serializersModule = SimplePolymorphicModule }
 
     @Test
     fun testSealedWithOneSubclass() {
