+++ formats/cbor/jvmTest/src/kotlinx/serialization/cbor/CborWriterSpecTest.kt
@@ -13,9 +13,9 @@ import kotlinx.serialization.cbor.internal.*
 class CborWriterSpecTest : WordSpec() {
     init {
 
-        fun withEncoder(block: Cbor.CborEncoder.() -> Unit): String {
+        fun withEncoder(block: CborEncoder.() -> Unit): String {
             val result = ByteArrayOutput()
-            Cbor.CborEncoder(result).block()
+            CborEncoder(result).block()
             return HexConverter.printHexBinary(result.toByteArray()).toLowerCase()
         }
